Move columns/utils.h into columns/column.h
The Wrap-related helpers (SliceVector, HasWrapMethod, WrapColumn) depend
only on declarations already provided by column.h, so fold them directly
into column.h and drop the separate utils.h header. Since column.h is
included everywhere, these utilities are now available without an extra
include.

Remove the now-redundant #include "utils.h" from the column sources and
add a warning message when attempting to import "columns/utils.h"
