+++ clickhouse/columns/enum.cpp
@@ -1,5 +1,4 @@
 #include "enum.h"
-#include "utils.h"
 
 #include "../base/input.h"
 #include "../base/output.h"
