+++ clickhouse/columns/string.cpp
@@ -1,5 +1,4 @@
 #include "string.h"
-#include "utils.h"
 
 #include "../base/wire_format.h"
